+++ src/main/java/walkingkooka/validation/expression/ValidatorExpressionEvaluationContextTesting2.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 Miroslav Pokorny (github.com/mP1)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package walkingkooka.validation.expression;
+
+import walkingkooka.tree.expression.ExpressionEvaluationContextTesting2;
+import walkingkooka.validation.ValidationReference;
+import walkingkooka.validation.form.HasFormTesting;
+
+public interface ValidatorExpressionEvaluationContextTesting2<R extends ValidationReference, C extends ValidatorExpressionEvaluationContext<R>> extends ValidatorExpressionEvaluationContextTesting,
+    ExpressionEvaluationContextTesting2<C>,
+    HasFormTesting<R> {
+
+    // ExpressionEvaluationContextTesting...............................................................................
+
+    @Override
+    default C createConverterLike() {
+        return this.createContext();
+    }
+
+    @Override
+    default String typeNameSuffix() {
+        return ValidatorExpressionEvaluationContext.class.getSimpleName();
+    }
+}
